Use macrotasks instead of microtasks to execute sql.
If the polling for new queries is executed as a microtask, it causes a
starvation of other macrotasks, for example XHR requests. By means of
setImmediate(), each SQL command is executed as a macrotask which let
some 'space' available to other tasks to be executed.

Reuse of already open connection if any, in getVersion() method.

WR-Issue: #12856